requirements.txt
django-allauth==0.40.0
django-debug-toolbar==2.0
django-filter==2.2.0
django-rest-auth==0.9.5
Django==2.2.4
djangorestframework==3.10.2
factory-boy==2.12.0
Faker==2.0.1
Markdown==3.1.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pytest-factoryboy==2.0.3
pytest==5.1.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.7.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2